images now rendering, loader organized into seperate file
