--- src/schemas/image.ts
@@ -1,9 +1,0 @@
-import { z } from 'astro:content'
-
-export default z
-  .object({
-    alt: z.string(),
-  })
-  .partial()
-  .passthrough()
-  .optional()
